Internal chat bug fix: Startup
On startup, an internal chat would fail due to trying to show that there is someone else talking when there is someone else on that internal chat, before the tkinter mainloop is initialised.

A try - except was added so that if the mainloop is not yet initialised, then it passes. Otherwise, runs normally.

Why try? Firstly, a try is faster than an if, as long as the except clause goes relatively unused (which here it does), and secondly, there is no way in tkinter to test if you're in the mainloop, and I wasn't about to extend the tkinter class.

Additionally, added a check that the self.current_target is not None to ensure that... it's not None when called.